Combined the index and character pages into one and updated the pop up box for the lessson
-Now minigame 1 has only the index page, there isn't two separate pages for the tutorial and minigame
-added the pop up lesson
