+++ src/minigames/minigame2-budgeting/lessons.json
@@ -1,0 +1,34 @@
+{
+    "lessons": [
+        {
+            "id": 0,
+            "lessonContent": "Welcome to the budgeting mini-game! Before playing the mini-game, we are going to discuss budgeting and making a plan for your money."
+        },
+
+        {
+            "id": 1,
+            "lessonContent": "Both money and time are limited resources, so you need to be careful about how you use them. That's why it is important to make a plan to manage them."
+        },
+
+        {
+            "id": 2,
+            "lessonContent": "Budgeting is about managing those resources and making a plan for how you will use those resources. With a budget, you will make a plan for how you will spend your money on the different things you need and want. This plan will also help you save for long term goals. You might want to go to college, buy a car, or save for a home. All these things cost a lot of money, and it takes time to earn money. You also don't want to be working all the time. You need time to rest and take care of your health. You also want to have time to spend with family and friends. A budget will help you stay on track and make wise decisions about using your money and time."
+        },
+
+        {
+            "id": 3,
+            "lessonContent": "There are many tools online to help you make a budget, but you can also just start by using a piece of paper or a spreadsheet. Write down how much you earn in one month. This can be money from your job, allowance, or gifts. The money you earn is your income. Then write down how much you spend each month on both things you want and things you need. Look at the difference between how much you are spending and  how much you are earning each month. This gives you an idea of how much money you are saving. If you are spending more money than you are earning, you might need to adjust how much you are spending, so that you can save for long term goals and avoid debt. "
+        },
+
+        {
+            "id": 4,
+            "lessonContent": "We won't go into much more detail about creating a budget, but this is how you can start to make one and understand your finances."
+        },
+
+        {
+            "id": 5,
+            "lessonContent": "Most of this information about budgeting  is from  khan academy, so if you want more information or want to review, check out their resources."
+
+        }
+    ]
+}
