+++ src/types/General.ts
@@ -3,3 +3,14 @@ export interface PopupProps {
   content: string;
   onClose: () => void;
 }
+
+export interface PopupPropsLesson {
+  title: string;
+  contentID: number;
+  onClickNext: () => void;
+}
+
+export interface lessonData {
+    id: number;
+    lessonContent: string;
+}
